requirements.txt
torch
torchvision
transformers>=4.45.0
accelerate
timm
opencv-contrib-python
scipy
shapely
trimesh[easy]
numpy-stl
fastapi[standard]
python-multipart
Pillow
aiofiles
aiosqlite
pydantic-settings
google-genai
openai
pillow-heif
bambulabs_api
